update ece docs (#38)
- feat: replace GEC references in ECE documentation with WIIT
- refactor: update ece screenshots
The old ones were inconsistent and partially only readable on ultra wide
screens. Also they were outdated.
Additonally removed the last screenshot for ILM docs, because it did not
provide any valuable additional
information. The popup that was show in the screenshot wil be presented
to the user anyways and is self-expainatory.
    Without any additonal text, it helps nobody there.